test: move TestFixture object to model
