+++ src/test/kotlin/blackjack/model/holder/HandTest.kt
@@ -1,5 +1,6 @@
 package blackjack.model.holder
 
+import blackjack.model.TestFixture
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
